+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/_params.py
@@ -9,7 +9,11 @@ import platform
 
 from argcomplete.completers import FilesCompleter
 
-from azure.cli.core.commands import register_cli_argument, CliArgumentType, register_extra_cli_argument
+from azure.cli.core.commands import (
+    CliArgumentType,
+    register_cli_argument,
+    register_extra_cli_argument)
+from azure.cli.core.commands.parameters import tags_type
 from azure.cli.core.commands.parameters import (
     enum_choice_list,
     file_type,
@@ -55,8 +59,10 @@ def _get_default_install_location(exe_name):
 
 name_arg_type = CliArgumentType(options_list=('--name', '-n'), metavar='NAME')
 
+register_cli_argument('acs', 'tags', tags_type)
+
 register_cli_argument('acs', 'name', arg_type=name_arg_type, configured_default='acs',
-                      help="ACS cluster name. You can configure the default using 'az configure --defaults acs=<name>'",
+                      help="ACS cluster name. You can configure the default using `az configure --defaults acs=<name>`",
                       completer=get_resource_name_completion_list('Microsoft.ContainerService/ContainerServices'))
 
 register_cli_argument('acs', 'resource_group', arg_type=resource_group_name_type)