+++ src/command_modules/azure-cli-storage/tests/storage_test_util.py
@@ -27,3 +27,8 @@ class StorageScenarioMixin(object):
         container_name = self.create_random_name(prefix=prefix, length=length)
         self.storage_cmd('storage container create -n {}', account_info, container_name)
         return container_name
+
+    def create_share(self, account_info, prefix='share', length=24):
+        share_name = self.create_random_name(prefix=prefix, length=length)
+        self.storage_cmd('storage share create -n {}', account_info, share_name)
+        return share_name