+++ src/command_modules/azure-cli-keyvault/HISTORY.rst
@@ -3,6 +3,14 @@
 Release History
 ===============
 
+unreleased
+++++++++++++++++++++
+
+* BC:`az keyvault certificate download` change -e from string or binary to PEM or DER to better represent the options
+* BC: Remove --expires and --not-before from `keyvault certificate create` as these parameters are not supported by the service.
+* Adds the --validity parameter to `keyvault certificate create` to selectively override the value in --policy
+* Fixes issue in `keyvault certificate get-default-policy` where 'expires' and 'not_before' were exposed but 'validity_in_months' was not.
+
 2.0.1 (2017-04-17)
 ++++++++++++++++++++